# packages/docs/src/openapi/eliza-v1.yaml
openapi: 3.0.0
info:
  title: alan OS API
  description: |-
    API documentation for alan OS v1.0.10 - A flexible and scalable AI agent framework.

    This API is designed to be used with a locally running alan instance. Endpoints allow for creating,
    managing, and interacting with AI agents through a REST interface.

    The API is organized into the following domains:
    - **System**: System-wide operations and environment management
    - **Agents**: Agent lifecycle and management operations
    - **Memory**: Agent memory and room management
    - **Messaging**: Message handling, channels, and servers
    - **Audio**: Audio processing and speech synthesis
    - **Media**: File upload and media management
    - **TEE**: Trusted Execution Environment operations
    - **WebSocket**: Real-time communication via Socket.IO
  version: 1.0.10
  contact:
    name: alan OS Community
    url: https://github.com/alanos/alan
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /api/audio/{agentId}/audio-messages/synthesize:
    post:
      tags:
        - audio
      summary: Synthesize speech from text
      description: Convert text to speech using agent's voice settings
      operationId: synthesizeSpeech
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Text to convert to speech
                options:
                  type: object
                  properties:
                    voice:
                      type: string
                      description: Voice ID or name
                    language:
                      type: string
                      description: Language code
      responses:
        '200':
          description: Speech synthesized successfully
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error synthesizing speech
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/audio/{agentId}/speech/generate:
    post:
      tags:
        - audio
      summary: Generate speech from text
      description: Generate speech audio from text using agent's voice settings
      operationId: generateSpeech
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID of the agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Text to convert to speech
      responses:
        '200':
          description: Speech generated successfully
          content:
            audio/mpeg:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Error generating speech
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
            message:
              type: string
              description: Error message
            details:
              type: string
              description: Detailed error information
...